Modify: 컨벤션에 따른 파일위치 및 코드 수정
