+++ next-app/src/components/common/Nav/Nav.tsx
@@ -1,14 +1,13 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
 
-import { TopNav } from './TopNav'
-import { SideMenuBar } from './SideNavBar/SideMenuBar'
+import { TopNavBar } from '../TopNavBar/TopNavBar'
+import { SideMenuBar } from '../SideNavBar/SideMenuBar'
 
 export const Nav = () => {
   const [showSideBar, setShowSideBar] = useState(false)
   return (
     <>
-      <TopNav openSideBar={() => setShowSideBar(true)} />
+      <TopNavBar openSideBar={() => setShowSideBar(true)} />
       {showSideBar && (
         <SideMenuBar closeSideBar={() => setShowSideBar(false)} />
       )}
