+++ next-app/src/components/common/SideNavBar/MenuItem/MenuItem.style.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+export const MenuItemWrap = styled.div`
+  padding: 12px 40px;
+  display: flex;
+  align-items: center;
+`
+
+export const MenuButton = styled.button`
+  font-size: 14px;
+  font-weight: 600;
+  color: #8c8c8c;
+  display: flex;
+  border: 0;
+  background: none;
+  cursor: pointer;
+`
