+++ next-app/src/components/common/TopNavBar/TopNavBar.style.ts
@@ -1,0 +1,37 @@
+import styled from "styled-components"
+
+export const TopNavContainer = styled.div`
+  width: 100%;
+  position: fixed;
+  height: 60px;
+  padding: 18px 40px;
+  background-color: #212322;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`
+
+export const MenuButton = styled.button`
+  border: 0;
+  background: none;
+  color: white;
+  font-size: 14px;
+  font-weight: 600;
+  display: flex;
+  cursor: pointer;
+`
+
+export const Feedoong = styled.span`
+  font-size: 20px;
+  font-weight: 700;
+  color: white;
+`
+
+export const MyPageButton = styled.div`
+  display: flex;
+  align-items: center;
+  font-size: 14px;
+  font-weight: 600;
+  color: white;
+  cursor: pointer;
+`
